requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.3
django-appconf==1.0.6
django-compressor==4.5.1
django-environ==0.11.2
django-filter==24.3
django-smtp-ssl==1.0
djangorestframework==3.15.2
djongo==1.3.7
dnspython==2.7.0
et_xmlfile==2.0.0
fuzzywuzzy==0.18.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
mongoengine==0.29.1
msgpack==1.1.0
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pymongo==3.11.4
pyOpenSSL==24.2.1
python-decouple==3.8
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
RapidFuzz==3.10.1
rcssmin==1.1.2
redis==5.2.0
rfc3986==1.5.0
rjsmin==1.2.2
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
uvicorn==0.32.0
whitenoise==6.8.2
zope.interface==7.1.1
